+++ ICD.Connect.Conferencing.Zoom/Components/Call/ZoomParticipant.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using ICD.Common.Utils;
-using ICD.Common.Utils.EventArguments;
-using ICD.Common.Utils.Extensions;
-using ICD.Connect.API.Commands;
+﻿using ICD.Common.Utils;
 using ICD.Connect.API.Nodes;
 using ICD.Connect.Conferencing.EventArguments;
 using ICD.Connect.Conferencing.Participants;
@@ -11,93 +6,12 @@ using ICD.Connect.Conferencing.Zoom.Responses;
 
 namespace ICD.Connect.Conferencing.Zoom.Components.Call
 {
-	public sealed class ZoomParticipant : IWebParticipant
+	public sealed class ZoomParticipant : AbstractWebParticipant
 	{
-		/// <summary>
-		/// Raised when the participant's status changes.
-		/// </summary>
-		public event EventHandler<ParticipantStatusEventArgs> OnStatusChanged;
-
-		/// <summary>
-		/// Raised when the participant's source type changes.
-		/// </summary>
-		public event EventHandler<ParticipantTypeEventArgs> OnSourceTypeChanged;
-
-		/// <summary>
-		/// Raised when the participant's name changes.
-		/// </summary>
-		public event EventHandler<StringEventArgs> OnNameChanged;
-
-		/// <summary>
-		/// Raised when the participant's mute state changes.
-		/// </summary>
-		public event EventHandler<BoolEventArgs> OnIsMutedChanged;
-
-		private string m_Name;
-		private eCallType m_SourceType;
-		private eParticipantStatus m_Status;
-		private bool m_IsMuted;
 		private readonly ZoomRoom m_ZoomRoom;
-
-		#region Properties
-
-		public string Name
-		{
-			get { return m_Name; }
-			private set
-			{
-				if (m_Name == value)
-					return;
-
-				m_Name = value;
-				OnNameChanged.Raise(this, new StringEventArgs(m_Name));
-			}
-		}
-
-		public eCallType SourceType
-		{
-			get { return m_SourceType; }
-			private set
-			{
-				if (m_SourceType == value)
-					return;
-				m_SourceType = value;
-				OnSourceTypeChanged.Raise(this, new ParticipantTypeEventArgs(m_SourceType));
-			}
-		}
-
-		public eParticipantStatus Status
-		{
-			get { return m_Status; }
-			private set
-			{
-				if (m_Status == value)
-					return;
-
-				m_Status = value;
-				OnStatusChanged.Raise(this, new ParticipantStatusEventArgs(m_Status));
-			}
-		}
-
-		public bool IsMuted
-		{
-			get { return m_IsMuted; }
-			private set
-			{
-				if (m_IsMuted == value)
-					return;
-
-				m_IsMuted = value;
-				OnIsMutedChanged.Raise(this, new BoolEventArgs(m_IsMuted));
-			}
-		}
 
 		public string UserId { get; private set; }
 		public string AvatarUrl { get; private set; }
-		public DateTime? Start { get; private set; }
-		public DateTime? End { get; private set; }
-
-		#endregion
 
 		public ZoomParticipant(ZoomRoom zoomRoom, ParticipantInfo info)
 		{
@@ -115,15 +29,17 @@ namespace ICD.Connect.Conferencing.Zoom.Components.Call
 			Status = eParticipantStatus.Connected;
 			SourceType = info.IsSendingVideo ? eCallType.Video : eCallType.Audio;
 			IsMuted = info.AudioState == eAudioState.AUDIO_MUTED;
+			IsHost = info.IsHost;
+			IsSelf = info.IsMyself;
 			AvatarUrl = info.AvatarUrl;
 		}
 
-		public void Kick()
+		public override void Kick()
 		{
 			m_ZoomRoom.SendCommand("zCommand Call Expel Id: {0}", UserId);
 		}
 
-		public void Mute(bool mute)
+		public override void Mute(bool mute)
 		{
 			m_ZoomRoom.SendCommand("zCommand Call MuteParticipant mute: {0} Id: {1}", 
 				mute ? "on" : "off", 
@@ -132,32 +48,29 @@ namespace ICD.Connect.Conferencing.Zoom.Components.Call
 
 		#endregion
 
+		#region Private Methods
+
+		// internal only cause zoom sucks and this is the easiest way
+		// to set the correct host state when host changes
+		internal void SetIsHost(bool isHost)
+		{
+			IsHost = isHost;
+		}
+
+		#endregion
+
 		#region Console
 
-		public string ConsoleName { get { return Name; } }
+		public override string ConsoleHelp { get { return "Zoom conference participant"; } }
 
-		public string ConsoleHelp { get { return "Zoom conference participant"; } }
-
-		public IEnumerable<IConsoleNodeBase> GetConsoleNodes()
+		public override void BuildConsoleStatus(AddStatusRowDelegate addRow)
 		{
-			yield break;
+			base.BuildConsoleStatus(addRow);
+
+			addRow("User ID", UserId);
+			addRow("Avatar URL", AvatarUrl);
 		}
 
-		public void BuildConsoleStatus(AddStatusRowDelegate addRow)
-		{
-			addRow("Name", Name);
-			addRow("UserId", UserId);
-			addRow("Status", Status);
-			addRow("SourceType", SourceType);
-			addRow("Start", Start);
-			addRow("End", End);
-		}
-
-		public IEnumerable<IConsoleCommand> GetConsoleCommands()
-		{
-			yield return new GenericConsoleCommand<bool>("Mute", "Usage: Mute <true/false>", m => Mute(m));
-			yield return new ConsoleCommand("Kick", "Kick the participant", () => Kick());
-		}
 		#endregion
 	}
 }